Add files via upload
Modeling and Python programming for an online retailer who is planning the locations of 3 fulfillment centers that will be used over the next 20 years with some specific details.
This project is the final project for the course of Data Science for e-Retail and the Sharing Economy at Cornell Tech in spring of 2022 semester. 
My teammate in this project is Evan Teng.